防止精度丢失，修改为toBigDecimal

+++ ablibrary/src/main/java/cc/ab/base/ext/NumberExt.kt
@@ -34,23 +34,23 @@ fun Number.px2sp(): Int {
 //<editor-fold defaultstate="collapsed" desc="保留小数">
 //保留2位小数(根据第三位四舍五入)
 fun Number?.to2pointUp(): String {
-  return if (this == null) "0.00" else DecimalFormat("#0.00").also { it.roundingMode = RoundingMode.HALF_UP }.format(this.toDoubleMy())
+  return if (this == null) "0.00" else DecimalFormat("#0.00").also { it.roundingMode = RoundingMode.HALF_UP }.format(this.toDoubleMy().toBigDecimal())
 }
 
 //保留2位小数(去掉后面的)
 fun Number?.to2pointDown(): String {
-  return if (this == null) "0.00" else DecimalFormat("#0.00").also { it.roundingMode = RoundingMode.DOWN }.format(this.toDoubleMy())
+  return if (this == null) "0.00" else DecimalFormat("#0.00").also { it.roundingMode = RoundingMode.DOWN }.format(this.toDoubleMy().toBigDecimal())
 }
 
 //保留2位小数+逗号分隔
 fun Number?.to2point2(): String {
-  return if (this == null) "0.00" else DecimalFormat("###,###,###,###,##0.00").format(this.toDoubleMy())
+  return if (this == null) "0.00" else DecimalFormat("###,###,###,###,##0.00").format(this.toDoubleMy().toBigDecimal())
 }
 
 //保留2位小数
 fun String?.to2point(): String {
   return if (this.isNullOrBlank()) "0.00" else try {
-    DecimalFormat("#0.00").format(this.toDoubleMy())
+    DecimalFormat("#0.00").format(this.toDoubleMy().toBigDecimal())
   } catch (e: Exception) {
     this
   }
@@ -59,7 +59,7 @@ fun String?.to2point(): String {
 //保留2位小数+逗号分隔
 fun String?.to2point2(): String {
   return if (this.isNullOrBlank()) "0.00" else try {
-    DecimalFormat("###,###,###,###,##0.00").format(this.toDoubleMy())
+    DecimalFormat("###,###,###,###,##0.00").format(this.toDoubleMy().toBigDecimal())
   } catch (e: Exception) {
     this
   }
@@ -76,20 +76,20 @@ fun Number?.toStart0(): String {
 
 //删除默尾0
 fun Number?.delEnd0(): String {
-  return if (this == null) "0" else DecimalFormat("#########0.##########").format(this.toDoubleMy())
+  return if (this == null) "0" else DecimalFormat("#########0.##########").format(this.toDoubleMy().toBigDecimal())
 }
 //</editor-fold>
 
 //<editor-fold defaultstate="collapsed" desc="保证位数">
 //2位数(0是1位)
 fun Int?.to2Num(): String {
-  return if (this == null || this.toDoubleMy() == 0.0) "0" else DecimalFormat("00").format(this.toDoubleMy())
+  return if (this == null || this.toDoubleMy() == 0.0) "0" else DecimalFormat("00").format(this.toDoubleMy().toBigDecimal())
 }
 //</editor-fold>
 
 //2位数
 fun Int?.to2Num2(): String {
-  return if (this == null || this.toDoubleMy() == 0.0) "00" else DecimalFormat("00").format(this.toDoubleMy())
+  return if (this == null || this.toDoubleMy() == 0.0) "00" else DecimalFormat("00").format(this.toDoubleMy().toBigDecimal())
 }
 //</editor-fold>
